pyproject.toml
[tool.poetry]
name = "sr-rules"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "WTFPL"
readme = "README.md"
packages = [{include = "sr_rules"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
